Removed tests.utils, moved functions to test_png
